pyproject.toml
[project]
name = "python-pytest-tdd"
version = "0.1.0"
description = ""
authors = [
    {name = "wupengyue"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "pytest-playwright (>=0.7.2,<0.8.0)",
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "lxml (>=6.0.2,<7.0.0)"
]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
